refactor: kapsle fleet telemetry konfiguration
