+++ main.js
@@ -2,16 +2,14 @@
 
 const utils = require('@iobroker/adapter-core');
 const axios = require('axios').default;
-const https = require('node:https');
 const qs = require('qs');
 const Json2iob = require('./lib/json2iob');
 const { loadProtos, TeslaCommandSigner, parseECKeyFromPem } = require('./lib/teslaSign');
+const { FleetTelemetryManager } = require('./lib/fleetTelemetry');
 const {
-  FleetTelemetryManager,
-  LOCATION_SCOPE_TELEMETRY_FIELDS,
-  buildFleetTelemetryProxyPayload,
-  parseTelemetryFieldsConfig,
-} = require('./lib/fleetTelemetry');
+  FleetTelemetryConfigurationManager,
+  isFleetTelemetryAdminCommand,
+} = require('./lib/fleetTelemetryConfig');
 
 // Fleet API regional endpoints (like HA const.py:17-18)
 const FLEET_API_REGIONS = {
@@ -57,7 +55,7 @@ class Teslamotors extends utils.Adapter {
 
     this.requestClient = axios.create();
     this.telemetryManager = null;
-    this.telemetryProxyClient = null;
+    this.telemetryConfigurationManager = new FleetTelemetryConfigurationManager(this);
   }
 
   /**
@@ -188,102 +186,9 @@ class Teslamotors extends utils.Adapter {
     );
   }
 
-  /**
-   * Returns all currently known vehicle VINs from the loaded product list.
-   *
-   * @returns {string[]}
-   */
-  getKnownVehicleVins() {
-    return this.idArray.filter((product) => product.type === 'vehicle').map((product) => product.vin);
-  }
 
-  /**
-   * Lazily builds an Axios client for the local vehicle-command proxy.
-   */
-  getTelemetryProxyClient() {
-    if (this.telemetryProxyClient) {
-      return this.telemetryProxyClient;
-    }
 
-    const baseURL = String(this.config.telemetryProxyUrl || '').trim().replace(/\/+$/, '');
-    if (!baseURL) {
-      throw new Error('telemetryProxyUrl is not configured');
-    }
 
-    this.telemetryProxyClient = axios.create({
-      baseURL,
-      timeout: 30000,
-      httpsAgent: new https.Agent({
-        rejectUnauthorized: !this.config.telemetryProxyAllowInsecure,
-      }),
-      headers: {
-        Accept: 'application/json',
-        'Content-Type': 'application/json',
-      },
-    });
-    return this.telemetryProxyClient;
-  }
-
-  /**
-   * Creates telemetry related info states used by the MQTT bridge and admin actions.
-   */
-  async createTelemetryInfoStates() {
-    const states = [
-      {
-        id: 'info.telemetryConnected',
-        common: { name: 'Telemetry MQTT connected', type: 'boolean', role: 'indicator.connected', read: true, write: false, def: false },
-      },
-      {
-        id: 'info.telemetryConfigured',
-        common: { name: 'Telemetry configured', type: 'boolean', role: 'indicator', read: true, write: false, def: false },
-      },
-      {
-        id: 'info.telemetrySynced',
-        common: { name: 'Telemetry synced on vehicle', type: 'boolean', role: 'indicator', read: true, write: false, def: false },
-      },
-      {
-        id: 'info.telemetryLastMessage',
-        common: { name: 'Last telemetry MQTT message metadata', type: 'string', role: 'json', read: true, write: false, def: '' },
-      },
-      {
-        id: 'info.telemetryLastError',
-        common: { name: 'Last telemetry error', type: 'string', role: 'text', read: true, write: false, def: '' },
-      },
-      {
-        id: 'info.telemetryLastApiSync',
-        common: { name: 'Last periodic Fleet API sync run', type: 'string', role: 'json', read: true, write: false, def: '' },
-      },
-      {
-        id: 'info.telemetryLastVehicleDataSync',
-        common: { name: 'Last vehicle_data API sync run', type: 'string', role: 'json', read: true, write: false, def: '' },
-      },
-      {
-        id: 'info.telemetryLastChargeHistorySync',
-        common: { name: 'Last charge history API sync run', type: 'string', role: 'json', read: true, write: false, def: '' },
-      },
-    ];
-
-    for (const state of states) {
-      await this.setObjectNotExistsAsync(state.id, {
-        type: 'state',
-        common: state.common,
-        native: {},
-      });
-    }
-  }
-
-  /**
-   * Normalizes Tesla/Fleet API response objects.
-   *
-   * @param {any} data
-   * @returns {any}
-   */
-  extractTeslaResponse(data) {
-    if (!data) {
-      return null;
-    }
-    return data.response !== undefined ? data.response : data;
-  }
 
   /**
    * Builds a user-facing error message and appends Tesla/Fleet API details when
@@ -311,73 +216,6 @@ class Teslamotors extends utils.Adapter {
     return message;
   }
 
-  /**
-   * Normalizes fleet status responses to a VIN keyed map.
-   *
-   * @param {any} data
-   * @returns {Record<string, any>}
-   */
-  normalizeFleetStatusResponse(data) {
-    const response = this.extractTeslaResponse(data);
-    if (!response) {
-      return {};
-    }
-
-    if (Array.isArray(response)) {
-      return response.reduce((accumulator, entry) => {
-        const vin = entry.vin || entry.VIN;
-        if (vin) {
-          accumulator[vin] = entry;
-        }
-        return accumulator;
-      }, {});
-    }
-
-    if (response.vehicles && Array.isArray(response.vehicles)) {
-      return response.vehicles.reduce((accumulator, entry) => {
-        const vin = entry.vin || entry.VIN;
-        if (vin) {
-          accumulator[vin] = entry;
-        }
-        return accumulator;
-      }, {});
-    }
-
-    if (response.vehicle_info && typeof response.vehicle_info === 'object' && !Array.isArray(response.vehicle_info)) {
-      const keyPairedVins = new Set(response.key_paired_vins || response.keyPairedVins || []);
-      const unpairedVins = new Set(response.unpaired_vins || response.unpairedVins || []);
-
-      return Object.entries(response.vehicle_info).reduce((accumulator, [vin, entry]) => {
-        if (!vin) {
-          return accumulator;
-        }
-
-        const status = {
-          vin,
-          ...(entry && typeof entry === 'object' && !Array.isArray(entry) ? entry : {}),
-        };
-
-        if (keyPairedVins.has(vin)) {
-          status.key_paired = true;
-        } else if (unpairedVins.has(vin)) {
-          status.key_paired = false;
-        }
-
-        accumulator[vin] = status;
-        return accumulator;
-      }, {});
-    }
-
-    if (typeof response === 'object') {
-      const keys = Object.keys(response);
-      const looksLikeVinMap = keys.every((key) => /^[A-HJ-NPR-Z0-9]{10,}$/.test(key));
-      if (looksLikeVinMap) {
-        return response;
-      }
-    }
-
-    return {};
-  }
 
   /**
    * Decode JWT token without signature verification (like HA __init__.py:81).
@@ -414,15 +252,8 @@ class Teslamotors extends utils.Adapter {
 
   async onReady() {
     this.setState('info.connection', false, true);
-    await this.createTelemetryInfoStates();
-    await this.setStateAsync('info.telemetryConnected', false, true);
-    await this.setStateAsync('info.telemetryConfigured', false, true);
-    await this.setStateAsync('info.telemetrySynced', false, true);
-    await this.setStateAsync('info.telemetryLastMessage', '', true);
-    await this.setStateAsync('info.telemetryLastError', '', true);
-    await this.setStateAsync('info.telemetryLastApiSync', '', true);
-    await this.setStateAsync('info.telemetryLastVehicleDataSync', '', true);
-    await this.setStateAsync('info.telemetryLastChargeHistorySync', '', true);
+    await this.telemetryConfigurationManager.createInfoStates();
+    await this.telemetryConfigurationManager.resetInfoStates();
 
     // Load protobuf definitions for command signing
     try {
@@ -499,7 +330,7 @@ class Teslamotors extends utils.Adapter {
         this.telemetryManager = new FleetTelemetryManager(this);
         await this.telemetryManager.start();
         try {
-          await this.refreshTelemetryConfigurationStates();
+          await this.telemetryConfigurationManager.refreshConfigurationStates();
         } catch (error) {
           this.log.warn('Could not refresh Fleet Telemetry configuration state: ' + error.message);
           await this.setStateAsync('info.telemetryLastError', error.message, true);
@@ -909,334 +740,16 @@ class Teslamotors extends utils.Adapter {
     return anySuccess;
   }
 
-  /**
-   * Parses Tesla firmware versions such as 2025.2.6 into comparable number arrays.
-   *
-   * @param {string} version
-   * @returns {number[]}
-   */
-  parseFirmwareVersion(version) {
-    return String(version || '')
-      .split('.')
-      .map((part) => parseInt(part, 10))
-      .filter((part) => !isNaN(part));
-  }
 
-  /**
-   * Returns true when the firmware is equal or newer than the provided minimum.
-   *
-   * @param {string} version
-   * @param {number[]} minimum
-   * @returns {boolean}
-   */
-  isMinimumFirmwareVersion(version, minimum) {
-    const current = this.parseFirmwareVersion(version);
-    for (let index = 0; index < Math.max(current.length, minimum.length); index++) {
-      const left = current[index] || 0;
-      const right = minimum[index] || 0;
-      if (left > right) {
-        return true;
-      }
-      if (left < right) {
-        return false;
-      }
-    }
-    return true;
-  }
 
-  /**
-   * Returns true if the reported firmware version supports proxy based telemetry config.
-   *
-   * @param {string} version
-   * @returns {boolean}
-   */
-  isTelemetryFirmwareSupported(version) {
-    return this.isMinimumFirmwareVersion(version, [2024, 26]);
-  }
 
-  /**
-   * Attempts to derive the virtual-key pairing state from the fleet_status response.
-   *
-   * @param {any} status
-   * @returns {boolean|null}
-   */
-  getVirtualKeyStatus(status) {
-    const candidates = [
-      status.virtual_key_paired,
-      status.virtualKeyPaired,
-      status.application_key_paired,
-      status.applicationKeyPaired,
-      status.public_key_available,
-      status.publicKeyAvailable,
-      status.public_key_present,
-      status.publicKeyPresent,
-      status.public_key_paired,
-      status.publicKeyPaired,
-      status.key_paired,
-      status.keyPaired,
-    ].filter((value) => value !== undefined);
 
-    if (candidates.length === 0) {
-      return null;
-    }
-    return candidates.some((value) => value === true);
-  }
 
-  /**
-   * Executes a request against the local vehicle-command proxy with the current
-   * Fleet OAuth token.
-   *
-   * @param {import('axios').AxiosRequestConfig} requestConfig
-   * @returns {Promise<import('axios').AxiosResponse<any>>}
-   */
-  async requestTelemetryProxy(requestConfig) {
-    const client = this.getTelemetryProxyClient();
-    return client.request({
-      ...requestConfig,
-      headers: {
-        Authorization: 'Bearer ' + this.session.access_token,
-        ...(requestConfig.headers || {}),
-      },
-    });
-  }
 
-  /**
-   * Calls Tesla's fleet_status endpoint and returns both the raw response and a
-   * VIN keyed map to simplify validation in the admin actions.
-   *
-   * @param {string[]} [vins]
-   * @returns {Promise<{raw: any; normalized: Record<string, any>}>}
-   */
-  async getFleetStatus(vins = this.getKnownVehicleVins()) {
-    if (!vins.length) {
-      throw new Error('No Tesla vehicle found for fleet status');
-    }
 
-    const response = await this.requestClient({
-      method: 'post',
-      url: this.getFleetApiBaseUrl() + '/api/1/vehicles/fleet_status',
-      headers: this.getFleetHeaders(),
-      data: { vins },
-    });
 
-    return {
-      raw: response.data,
-      normalized: this.normalizeFleetStatusResponse(response.data),
-    };
-  }
 
-  /**
-   * Builds the standard telemetry configuration payload for the vehicle-command proxy.
-   *
-   * @param {string[]} vins
-   * @returns {{ vins: string[]; config: Record<string, any> }}
-   */
-  buildTelemetryConfigPayload(vins) {
-    if (!this.config.telemetryServerHost) {
-      throw new Error('telemetryServerHost is not configured');
-    }
-    if (!this.config.telemetryServerCaPem) {
-      throw new Error('telemetryServerCaPem is not configured');
-    }
 
-    const fields = parseTelemetryFieldsConfig(this.config.telemetryFieldsJson);
-    if (!this.scopes.includes('vehicle_location')) {
-      const omittedFields = [];
-      for (const fieldName of LOCATION_SCOPE_TELEMETRY_FIELDS) {
-        if (fields[fieldName]) {
-          delete fields[fieldName];
-          omittedFields.push(fieldName);
-        }
-      }
-      if (omittedFields.length) {
-        this.log.info(`vehicle_location scope missing, omit Fleet Telemetry location fields: ${omittedFields.join(', ')}`);
-      }
-    }
-
-    return buildFleetTelemetryProxyPayload(vins, {
-      hostname: this.config.telemetryServerHost,
-      port: this.config.telemetryServerPort,
-      ca: this.config.telemetryServerCaPem,
-      fields,
-    });
-  }
-
-  /**
-   * Reads telemetry configuration state for a VIN.
-   *
-   * @param {string} vin
-   * @returns {Promise<any>}
-   */
-  async getFleetTelemetryConfig(vin) {
-    const response = await this.requestTelemetryProxy({
-      method: 'get',
-      url: `/api/1/vehicles/${vin}/fleet_telemetry_config`,
-    });
-    const data = this.extractTeslaResponse(response.data);
-    const configured = !!(data && data.config);
-
-    if (data && typeof data.synced === 'boolean') {
-      await this.setStateAsync('info.telemetrySynced', configured && data.synced === true, true);
-    }
-    await this.setStateAsync('info.telemetryConfigured', configured, true);
-    await this.setStateAsync('info.telemetryLastError', '', true);
-    return data;
-  }
-
-  /**
-   * Refreshes the diagnostic states that show whether Fleet Telemetry is already
-   * configured and synced on the known vehicles. This is intentionally called
-   * during startup too, so an already configured vehicle is reflected correctly
-   * after an adapter restart.
-   *
-   * @param {string[]} [vins]
-   * @returns {Promise<Record<string, any>>}
-   */
-  async refreshTelemetryConfigurationStates(vins = this.getKnownVehicleVins()) {
-    if (!vins.length) {
-      await this.setStateAsync('info.telemetryConfigured', false, true);
-      await this.setStateAsync('info.telemetrySynced', false, true);
-      return {};
-    }
-
-    const configs = {};
-    let anyConfigured = false;
-    let allConfiguredAreSynced = true;
-
-    for (const vin of vins) {
-      try {
-        configs[vin] = await this.getFleetTelemetryConfig(vin);
-        const configured = !!(configs[vin] && configs[vin].config);
-        anyConfigured = anyConfigured || configured;
-        if (configured && configs[vin].synced !== true) {
-          allConfiguredAreSynced = false;
-        }
-      } catch (error) {
-        allConfiguredAreSynced = false;
-        configs[vin] = { error: error.message };
-      }
-    }
-
-    await this.setStateAsync('info.telemetryConfigured', anyConfigured, true);
-    await this.setStateAsync('info.telemetrySynced', anyConfigured && allConfiguredAreSynced, true);
-    return configs;
-  }
-
-  /**
-   * Deletes a telemetry configuration from a VIN.
-   *
-   * @param {string} vin
-   * @returns {Promise<any>}
-   */
-  async deleteFleetTelemetryConfig(vin) {
-    const response = await this.requestTelemetryProxy({
-      method: 'delete',
-      url: `/api/1/vehicles/${vin}/fleet_telemetry_config`,
-    });
-    await this.setStateAsync('info.telemetryConfigured', false, true);
-    await this.setStateAsync('info.telemetrySynced', false, true);
-    await this.setStateAsync('info.telemetryLastError', '', true);
-    return this.extractTeslaResponse(response.data);
-  }
-
-  /**
-   * Validates the current vehicle status and configures Fleet Telemetry through
-   * the local vehicle-command proxy.
-   *
-   * @param {string[]} [vins]
-   * @returns {Promise<{status: any; configure: any; configs: Record<string, any>}>}
-   */
-  async configureFleetTelemetry(vins = this.getKnownVehicleVins()) {
-    if (!vins.length) {
-      throw new Error('No Tesla vehicle found for telemetry configuration');
-    }
-
-    const status = await this.getFleetStatus(vins);
-    /** @type {string[]} */
-    const validationErrors = [];
-
-    for (const vin of vins) {
-      const vehicleStatus = status.normalized[vin];
-      if (!vehicleStatus) {
-        validationErrors.push(`${vin}: fleet_status missing`);
-        continue;
-      }
-
-      const firmwareVersion = vehicleStatus.firmware_version || vehicleStatus.firmwareVersion;
-      const streamingToggleState = vehicleStatus.safety_screen_streaming_toggle_enabled;
-      if (typeof streamingToggleState === 'boolean') {
-        if (firmwareVersion && !this.isMinimumFirmwareVersion(firmwareVersion, [2025, 20])) {
-          validationErrors.push(`${vin}: unsupported_firmware (${firmwareVersion})`);
-        }
-        if (streamingToggleState === false) {
-          validationErrors.push(`${vin}: streaming_toggle_disabled`);
-        }
-      } else {
-        if (firmwareVersion && !this.isTelemetryFirmwareSupported(firmwareVersion)) {
-          validationErrors.push(`${vin}: unsupported_firmware (${firmwareVersion})`);
-        }
-
-        const virtualKeyState = this.getVirtualKeyStatus(vehicleStatus);
-        if (virtualKeyState === false) {
-          validationErrors.push(`${vin}: missing_key`);
-        }
-      }
-
-      if (vehicleStatus.limit_reached === true || vehicleStatus.config_limit_reached === true) {
-        validationErrors.push(`${vin}: max_configs`);
-      }
-    }
-
-    if (validationErrors.length > 0) {
-      const errorMessage = validationErrors.join('; ');
-      await this.setStateAsync('info.telemetryLastError', errorMessage, true);
-      throw new Error(errorMessage);
-    }
-
-    const payload = this.buildTelemetryConfigPayload(vins);
-    const configureResponse = await this.requestTelemetryProxy({
-      method: 'post',
-      url: '/api/1/vehicles/fleet_telemetry_config',
-      data: payload,
-    });
-
-    const configureData = this.extractTeslaResponse(configureResponse.data);
-    const skippedVehicles = Array.isArray(configureData && configureData.skipped_vehicles) ? configureData.skipped_vehicles : [];
-    if (skippedVehicles.length >= vins.length) {
-      const errorMessage = skippedVehicles
-        .map((entry) => `${entry.vin || entry.VIN || 'unknown'}: ${entry.reason || entry.code || 'skipped'}`)
-        .join('; ');
-      await this.setStateAsync('info.telemetryConfigured', false, true);
-      await this.setStateAsync('info.telemetrySynced', false, true);
-      await this.setStateAsync('info.telemetryLastError', errorMessage, true);
-      throw new Error(errorMessage);
-    }
-
-    const configs = {};
-    let allSynced = true;
-    for (const vin of vins) {
-      try {
-        configs[vin] = await this.getFleetTelemetryConfig(vin);
-        if (!configs[vin] || configs[vin].synced !== true) {
-          allSynced = false;
-        }
-      } catch (error) {
-        allSynced = false;
-        configs[vin] = { error: error.message };
-      }
-    }
-
-    await this.setStateAsync('info.telemetryConfigured', true, true);
-    await this.setStateAsync('info.telemetrySynced', allSynced, true);
-    await this.setStateAsync('info.telemetryLastError', '', true);
-
-    return {
-      status,
-      configure: configureData,
-      configs,
-      skippedVehicles,
-    };
-  }
 
   /**
    * Update all devices via Fleet API.
@@ -2208,44 +1721,9 @@ class Teslamotors extends utils.Adapter {
       return;
     }
 
-    if (
-      obj.command === 'checkFleetStatus' ||
-      obj.command === 'configureFleetTelemetry' ||
-      obj.command === 'getFleetTelemetryConfig' ||
-      obj.command === 'deleteFleetTelemetryConfig'
-    ) {
+    if (isFleetTelemetryAdminCommand(obj.command)) {
       try {
-        if (!this.session.access_token) {
-          throw new Error('Fleet API session is not available yet');
-        }
-        if (!this.idArray.length) {
-          await this.getDeviceList({ throwOnError: true });
-        }
-
-        const message = obj.message || {};
-        const vins = Array.isArray(message.vins)
-          ? message.vins
-          : message.vin
-            ? [message.vin]
-            : this.getKnownVehicleVins();
-
-        let result = null;
-        if (obj.command === 'checkFleetStatus') {
-          result = await this.getFleetStatus(vins);
-        } else if (obj.command === 'configureFleetTelemetry') {
-          result = await this.configureFleetTelemetry(vins);
-        } else if (obj.command === 'getFleetTelemetryConfig') {
-          result = {};
-          for (const vin of vins) {
-            result[vin] = await this.getFleetTelemetryConfig(vin);
-          }
-        } else if (obj.command === 'deleteFleetTelemetryConfig') {
-          result = {};
-          for (const vin of vins) {
-            result[vin] = await this.deleteFleetTelemetryConfig(vin);
-          }
-        }
-
+        const result = await this.telemetryConfigurationManager.handleAdminCommand(obj.command, obj.message || {});
         this.sendTo(obj.from, obj.command, { ok: true, result }, obj.callback);
       } catch (/** @type {any} */ error) {
         const errorMessage = this.extractErrorMessage(error);
